This PR adds some basic signal-processing functionality and assorted support code.

New `signalProcessing` library includes:
* FFT
* DCT
* IIR filters
* Linear- and mel-spaced filterbanks
* Hamming and "Hanning" windowing functions

New nodes:
* Nodes for the above signal processing functions
* BufferNode: outputs the concatenation of the most recent N inputs of the input

Misc:
* `RingBuffer` class (in utilities)
* Added stl-type iterators to `math::Vector`: allows using math vectors in stl algorithms and range-based for loops. (Implemented in `StlStridedIterator` in `utilities` library)
* Slight change in `wrap.py` to decouple module name from input model filename
* Moved the example models from `examples/data` to `examples/models`

*Notes*
* The FFT implementation in the `signalProcessing` library is pretty bare-bones.
* The compiled FFT implementation has a little bit of commented-out or unused code (alternative implementations) that may become useful when we have the autotuning optimizer.
* The DCT is implemented in a kind of stupid way that won't scale well for large N (but is performant for small N). It's OK for now, since we currently only use it on small feature vectors (typically of size 40).
* Some of the signal processing code uses math vectors, and some still uses `std::vector`. I'm not sure how objectionable that is right now.
* Some of the code is, I'm sure, not properly formatted. I'll try to fix that up soon. Seattle Public Utilities is about to shut off our power, so I may not get to it today.

Related work items: #1083, #1084, #1085, #1086, #1087, #811, #812

+++ interfaces/common/include/MathInterface.h
@@ -37,24 +37,23 @@ namespace math
         return static_cast<long>(index);
     }
 
-
     struct TensorShape
     {
-        size_t rows;
-        size_t columns;
-        size_t channels;
+        int rows;
+        int columns;
+        int channels;
 
         /// <summary> Gets the number of elements. </summary>
         ///
         /// <returns> The number of elements in the `TensorShape` </returns>
-        size_t Size() const { return rows * columns * channels; }
+        int Size() const { return rows * columns * channels; }
 
         TensorShape() : TensorShape(0, 0, 0)
         {
             // this is needed by SWIG wrappers.
         }
 
-        TensorShape(size_t rows, size_t columns, size_t channels) :
+        TensorShape(int rows, int columns, int channels) :
             rows(rows), columns(columns), channels(channels) 
         {
         }
@@ -76,7 +75,7 @@ namespace math
     class Tensor
     {
     public:
-        Tensor(const std::vector<ElementType>& data, size_t rows, size_t columns, size_t channels) :
+        Tensor(const std::vector<ElementType>& data, int rows, int columns, int channels) :
             shape{ rows, columns, channels }, data(data)
         {
         }
